fix: Price 필드 값 Int -> Long 타입 변경( Int 타입 최대값인 2,147,483,647의 범위 제한을 극복하기 위해 )

+++ app/src/main/java/com/mbj/ssassamarket/data/model/ProductPostItem.kt
@@ -14,7 +14,7 @@ data class ProductPostItem(
     @Json(name = "imageLocations")
     val imageLocations: List<String?>? = null,
     @Json(name = "price")
-    val price: Int,
+    val price: Long,
     @Json(name = "title")
     val title: String,
     @Json(name = "category")
@@ -35,7 +35,7 @@ data class ProductPostItem(
 
 data class PatchProductRequest(
     val title: String,
-    val price: Int,
+    val price: Long,
     val content: String
 )
